fix: 修复NRC与RPE坐标系比例尺不一致引起的父子线解绑问题 chore: 提高内部格式NRC的数值精度，防止工具计算时出现精度丢失 feat: 允许传入自定义比例尺用于转换到NRC与从NRC转换到其他格式 perf: 修改容差算法，在保持输出内容尽可能不变的同时，文件更小

+++ PhiFanmade.Core/PhiFanmadeNrc/Note.cs
@@ -35,7 +35,7 @@ namespace PhiFanmade.Core.PhiFanmadeNrc
         /// <summary>
         /// 音符相对于判定线的X坐标
         /// </summary>
-        public float PositionX = 0.0f; // X坐标
+        public double PositionX = 0.0f; // X坐标
 
         /// <summary>
         /// 音符宽度倍率
@@ -65,7 +65,7 @@ namespace PhiFanmade.Core.PhiFanmadeNrc
         /// <summary>
         /// 音符相对于判定线的Y轴偏移
         /// </summary>
-        public float YOffset = 0.0f; // Y偏移
+        public double YOffset = 0.0f; // Y偏移
 
         /// <summary>
         /// 音符颜色（RGB，顶点颜色乘法）
